refactor: extract RebuildHideMixin from governance storage
Move the RebuildHide bucket (hide_governance_agent_playbooks_for_rebuild,
apply_governance_agent_playbook_rebuild + the source-window / FTS-vec privates
and _build_agent_playbook_source_window_rows) out of the residual governance
mixin into a dedicated RebuildHideMixin, on SQLite and the storage_base ABC.
Verbatim move: apply_governance_agent_playbook_rebuild preserves its ordering
and every commit/rollback; _emit_hard_delete_playbook stays a local in-method
import, adjusted ._playbook -> .._playbook for the deeper governance/ package.

This is the final governance-decomposition extraction: the residual
SQLiteGovernanceMixin now holds 0 public governance-contract methods (only the
conn/_lock/org_id declarations + the cross-bucket validators reached via MRO),
and the storage_base residual GovernanceMixin ABC is fully drained.
