+++ reflexio/server/services/storage/sqlite_storage/governance/__init__.py
@@ -1,11 +1,13 @@
 from ._audit import AuditEventStoreMixin
 from ._erase_execution import GovernanceEraseExecutionMixin
 from ._purge import PurgeOperationStoreMixin
+from ._rebuild_hide import RebuildHideMixin
 from ._subject_barrier import SubjectBarrierMixin
 
 __all__ = [
     "AuditEventStoreMixin",
     "GovernanceEraseExecutionMixin",
     "PurgeOperationStoreMixin",
+    "RebuildHideMixin",
     "SubjectBarrierMixin",
 ]
